fix: a tokio test that starved its own timer, and the 500-line cap
Two failures from the Windows job.

**The relay-timeout test was not flaky, it was starved.**
`SlowThenBusy::launch` blocks its worker with `std::thread::sleep`, and
`#[tokio::test(flavor = "multi_thread")]` sizes the pool from the CPU
count — so on a two-core runner the blocked launch can take the runtime
with it and the relay's own timeout timer never fires. The relay then
returned `Started` for a launch it was supposed to give up on, which is
the opposite of the property under test. Pinning `worker_threads = 4`
fixes the cause; widening the durations would only have hidden it, and
cost ten seconds a run to do so.

**`platform.rs` crossed the cap** at 533 lines. Its test module moves to
`platform/tests.rs` — the seam `service`, `identity` and `docker::command`
already use — rather than splitting the implementation, which has no
seam worth drawing: every function there is one platform question.
